# backend/docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TrippleShop",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/v1/orders/{order}/tracking": {
            "get": {
                "operationId": "order.tracking",
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "description": "The order ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "order_number": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderStatusHistory"
                                            }
                                        },
                                        "shipment": {
                                            "$ref": "#/components/schemas/Shipment"
                                        }
                                    },
                                    "required": [
                                        "order_number",
                                        "status",
                                        "history",
                                        "shipment"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview.",
                                            "examples": [
                                                ""
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderStatusHistory": {
                "type": "array",
                "items": {
                    "type": "string"
                },
                "minItems": 0,
                "maxItems": 0,
                "additionalItems": false,
                "title": "OrderStatusHistory"
            },
...
            "Shipment": {
                "type": "array",
                "items": {
                    "type": "string"
                },
                "minItems": 0,
                "maxItems": 0,
                "additionalItems": false,
                "title": "Shipment"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}